requirements.txt
Flask==2.0.1
pandas==1.3.3
pywebview==3.5
gunicorn==20.1.0
